Extract loading permissions and action mapping logic to separate class
